Split integration test support
